feat(ngl theme): add more to ngl theme
