+++ packages/react-magma-dom/src/theme/magma/components/button.ts
@@ -1,3 +1,6 @@
+import { colors } from '../colors';
+import { darken } from 'polished';
+
 interface ThemeButton {
   backgroundColor: string;
   borderColor: string;
@@ -26,115 +29,123 @@ export interface ThemeButtons {
   disabled: ThemeButton;
 }
 
-export const button = colors => {
-  return {
-    primary: {
-      backgroundColor: colors.primary,
+export const button = {
+  primary: {
+    backgroundColor: colors.primary,
+    borderColor: colors.primary,
+    textColor: colors.neutral08,
+    inverse: {
+      backgroundColor: colors.neutral08,
+      borderColor: colors.neutral08,
+      textColor: colors.primary,
+    },
+    outline: {
       borderColor: colors.primary,
-      textColor: colors.neutral08,
+      textColor: colors.primary,
       inverse: {
-        backgroundColor: colors.neutral08,
         borderColor: colors.neutral08,
-        textColor: colors.primary,
-      },
-      outline: {
-        borderColor: colors.primary,
-        textColor: colors.primary,
-        inverse: {
-          borderColor: colors.neutral08,
-          textColor: colors.neutral08,
-        },
+        textColor: colors.neutral08,
       },
     },
-    secondary: {
+    hover: {
+      backgroundColor: darken(0.1, colors.primary),
+      borderColor: colors.primary,
+      textColor: colors.foundation,
+    },
+    active: {
+      backgroundColor: darken(0.2, colors.primary),
+      borderColor: colors.primary,
+      textColor: colors.foundation,
+    },
+  },
+  secondary: {
+    backgroundColor: colors.neutral08,
+    borderColor: colors.neutral05,
+    textColor: colors.neutral,
+    inverse: {
       backgroundColor: colors.neutral08,
+      borderColor: colors.neutral08,
+      textColor: colors.neutral,
+    },
+    outline: {
       borderColor: colors.neutral05,
       textColor: colors.neutral,
       inverse: {
-        backgroundColor: colors.neutral08,
         borderColor: colors.neutral08,
-        textColor: colors.neutral,
-      },
-      outline: {
-        borderColor: colors.neutral05,
-        textColor: colors.neutral,
-        inverse: {
-          borderColor: colors.neutral08,
-          textColor: colors.neutral08,
-        },
+        textColor: colors.neutral08,
       },
     },
-    success: {
-      backgroundColor: colors.success,
+  },
+  success: {
+    backgroundColor: colors.success,
+    borderColor: colors.success,
+    textColor: colors.neutral08,
+    inverse: {
+      backgroundColor: colors.neutral08,
+      borderColor: colors.neutral08,
+      textColor: colors.success,
+    },
+    outline: {
       borderColor: colors.success,
-      textColor: colors.neutral08,
+      textColor: colors.success,
       inverse: {
-        backgroundColor: colors.neutral08,
         borderColor: colors.neutral08,
-        textColor: colors.success,
-      },
-      outline: {
-        borderColor: colors.success,
-        textColor: colors.success,
-        inverse: {
-          borderColor: colors.neutral08,
-          textColor: colors.neutral08,
-        },
+        textColor: colors.neutral08,
       },
     },
-    danger: {
-      backgroundColor: colors.danger,
+  },
+  danger: {
+    backgroundColor: colors.danger,
+    borderColor: colors.danger,
+    textColor: colors.neutral08,
+    inverse: {
+      backgroundColor: colors.neutral08,
+      borderColor: colors.neutral08,
+      textColor: colors.danger,
+    },
+    outline: {
       borderColor: colors.danger,
-      textColor: colors.neutral08,
+      textColor: colors.danger,
       inverse: {
-        backgroundColor: colors.neutral08,
         borderColor: colors.neutral08,
-        textColor: colors.danger,
-      },
-      outline: {
-        borderColor: colors.danger,
-        textColor: colors.danger,
-        inverse: {
-          borderColor: colors.neutral08,
-          textColor: colors.neutral08,
-        },
+        textColor: colors.neutral08,
       },
     },
-    marketing: {
+  },
+  marketing: {
+    backgroundColor: colors.pop04,
+    borderColor: colors.pop04,
+    textColor: colors.foundation02,
+    inverse: {
       backgroundColor: colors.pop04,
       borderColor: colors.pop04,
       textColor: colors.foundation02,
+    },
+    outline: {
+      borderColor: colors.pop04,
+      textColor: colors.foundation02,
       inverse: {
-        backgroundColor: colors.pop04,
         borderColor: colors.pop04,
         textColor: colors.foundation02,
-      },
-      outline: {
-        borderColor: colors.pop04,
-        textColor: colors.foundation02,
-        inverse: {
-          borderColor: colors.pop04,
-          textColor: colors.foundation02,
-        },
       },
     },
-    disabled: {
+  },
+  disabled: {
+    backgroundColor: colors.neutral06,
+    borderColor: colors.neutral06,
+    textColor: colors.disabledText,
+    inverse: {
       backgroundColor: colors.neutral06,
       borderColor: colors.neutral06,
       textColor: colors.disabledText,
+    },
+    outline: {
+      borderColor: colors.disabledText,
+      textColor: colors.disabledText,
       inverse: {
-        backgroundColor: colors.neutral06,
-        borderColor: colors.neutral06,
-        textColor: colors.disabledText,
-      },
-      outline: {
-        borderColor: colors.disabledText,
-        textColor: colors.disabledText,
-        inverse: {
-          borderColor: colors.disabledInverseText,
-          textColor: colors.disabledInverseText,
-        },
+        borderColor: colors.disabledInverseText,
+        textColor: colors.disabledInverseText,
       },
     },
-  };
+  },
 };
